Add demo video processing flow
After the user uploads a demo video for processing,
we show them a countdown and a series of slides with information.
Once the countdown has finished,
we start to poll for the state of the video.
Once the video has finished processing,
we redirect the user to the videos page
so they can see the results.

* move onboarding slide images to main image folder
* Create OnboardingSlides to display slides we show to the user while their video is processing
* Create OnboardingSlide to encapsulate slide logic/display
* Create OnboardingCountdown to display estimated time remaining to user

https://neonlabs.atlassian.net/browse/NEON-1262

WIP Wire up email for video callback

* NOTE: This is a WIP as the endpoint for making a PUT request does not
seem to be working just yet.